requirements.txt
lxml >= 3.7.3
pyyaml >= 3.12
unirest >= 1.1.7
requests >= 2.14.2